Move Falcon encoded key helpers
