+++ crypto/internal/falcon1024/falcon1024.go
@@ -138,64 +138,6 @@ func initPrivateKey(priv *PrivateKey, f, g, ntruF, ntruG smallPolynomial, h ring
 	return priv, nil
 }
 
-// TestingOnlyNewPrivateKeyFromEncoded creates a private key from the Falcon
-// encoded private-key form, for testing purposes.
-//
-// Bytes must not be called on the resulting key, as it will return a
-// deterministic placeholder seed rather than the encoded private key.
-func TestingOnlyNewPrivateKeyFromEncoded(privBytes []byte) (*PrivateKey, error) {
-	f, g, ntruF, err := skDecode(privBytes)
-	if err != nil {
-		return nil, err
-	}
-
-	ntruG, ok := completePrivate(f, g, ntruF)
-	if !ok {
-		return nil, errors.New("falcon-1024: invalid private key")
-	}
-
-	h, ok := computePublic(f, g)
-	if !ok {
-		return nil, errors.New("falcon-1024: invalid private key")
-	}
-
-	priv := &PrivateKey{}
-	seed := sha3.NewSHAKE256()
-	_, _ = seed.Write(privBytes)
-	_, _ = seed.Read(priv.seed[:])
-
-	return initPrivateKey(priv, f, g, ntruF, ntruG, h)
-}
-
-// TestingOnlyNewPrivateKeyFromSeedWithEncoded returns the seed-backed private
-// key and Falcon encoded private-key bytes generated from seed, for testing
-// purposes.
-func TestingOnlyNewPrivateKeyFromSeedWithEncoded(seed []byte) (*PrivateKey, []byte, error) {
-	priv := &PrivateKey{}
-
-	if len(seed) != SeedSize {
-		return nil, nil, errInvalidSeedLength
-	}
-	copy(priv.seed[:], seed)
-
-	rng := sha3.NewSHAKE256()
-	_, _ = rng.Write(seed)
-
-	f, g, ntruF, ntruG, h := generateKeyComponents(rng)
-
-	privBytes := make([]byte, encodedPrivateKeySize)
-	if err := skEncode(privBytes, f, g, ntruF); err != nil {
-		return nil, nil, err
-	}
-
-	priv, err := initPrivateKey(priv, f, g, ntruF, ntruG, h)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	return priv, privBytes, nil
-}
-
 func newPublicKeyFromH(h ringElement) (PublicKey, error) {
 	var pub PublicKey
 	if err := pkEncode(pub.raw[:], h); err != nil {
